# fix deletion by key with leading space