@buger 2021-03-21
@daria-kay apparently OSS-Fuzz integration found issue with your PR. Would you like to check it?

https://google.github.io/oss-fuzz/advanced-topics/reproducing/

```

  | +----------------------------------------Release Build Stacktrace----------------------------------------+
-- | --
  | Command: /mnt/scratch0/clusterfuzz/bot/builds/clusterfuzz-builds_jsonparser_842b7742ab4fad0d63df92dcf84cec58705eaa50/revisions/fuzzdelete -rss_limit_mb=2560 -timeout=60 -runs=100 /mnt/scratch0/clusterfuzz/bot/inputs/fuzzer-testcases/crash-da39a3ee5e6b4b0d3255bfef95601890afd80709
  | Bot: oss-fuzz-linux-zone1-host-d2bp-0
  | Time ran: 0.035671234130859375
  |  
  | INFO: Running with entropic power schedule (0xFF, 100).
  | INFO: Seed: 1871214021
  | INFO: 2190 Extra Counters
  | /mnt/scratch0/clusterfuzz/bot/builds/clusterfuzz-builds_jsonparser_842b7742ab4fad0d63df92dcf84cec58705eaa50/revisions/fuzzdelete: Running 1 inputs 100 time(s) each.
  | Running: /mnt/scratch0/clusterfuzz/bot/inputs/fuzzer-testcases/crash-da39a3ee5e6b4b0d3255bfef95601890afd80709
  | panic: runtime error: index out of range [-1]
  |  
  | goroutine 17 [running, locked to thread]:
  | github.com/buger/jsonparser.findKeyStart(0x602000000090, 0x0, 0x0, 0x60fcb4, 0x4, 0x0, 0x0, 0x0)
  | github.com/buger/jsonparser/parser.go:64 +0xa6b
  | github.com/buger/jsonparser.Delete(0x602000000090, 0x0, 0x0, 0x10c000052e88, 0x1, 0x1, 0x55c1eb, 0x10c00002a060, 0x0)
  | github.com/buger/jsonparser/parser.go:731 +0xac6
  | github.com/buger/jsonparser.FuzzDelete(...)
  | github.com/buger/jsonparser/fuzz.go:31
  | main.LLVMFuzzerTestOneInput(...)
  | command-line-arguments/main.293696178.go:21
  | AddressSanitizer:DEADLYSIGNAL
  | =================================================================
  | ==15856==ERROR: AddressSanitizer: ABRT on unknown address 0x053900003df0 (pc 0x0000005bca01 bp 0x10c000052a48 sp 0x10c000052a30 T0)
  | SCARINESS: 10 (signal)
  | #0 0x5bca01 in runtime.raise runtime/sys_linux_amd64.s:164
  |  
  | AddressSanitizer can not provide additional info.
  | SUMMARY: AddressSanitizer: ABRT (/mnt/scratch0/clusterfuzz/bot/builds/clusterfuzz-builds_jsonparser_842b7742ab4fad0d63df92dcf84cec58705eaa50/revisions/fuzzdelete+0x5bca01)
  | ==15856==ABORTING
  |  
  |  
  | +----------------------------------------Release Build Unsymbolized Stacktrace (diff)----------------------------------------+
  |  
  | ==15856==ERROR: AddressSanitizer: ABRT on unknown address 0x053900003df0 (pc 0x0000005bca01 bp 0x10c000052a48 sp 0x10c000052a30 T0)
  | SCARINESS: 10 (signal)
  | #0 0x5bca01  (/mnt/scratch0/clusterfuzz/bot/builds/clusterfuzz-builds_jsonparser_842b7742ab4fad0d63df92dcf84cec58705eaa50/revisions/fuzzdelete+0x5bca01)
  |  
  | AddressSanitizer can not provide additional info.

```